281: create new /interface file structure
