+++ src/bot.js
@@ -3,7 +3,8 @@ import { PPCL, BRP } from './constants';
 import { getMapTraversabilityInCells } from './helpers/map';
 import { getMe } from './helpers/player';
 import { FSM } from './helpers/fsm';
-import { isAutonomousMode, isVisualMode, move, dequeueChatMessages } from './utils/interface';
+import { isAutonomousMode, isVisualMode, move } from './interface/keys';
+import { dequeueChatMessages } from './interface/chat';
 import { drawAllyPolypointPath } from './draw/triangulation';
 import { desiredAccelerationMultiplier } from './helpers/physics';
 import { getShortestPolypointPath } from './navmesh/path';
