+++ src/interface/test/chat.spec.js
@@ -2,17 +2,14 @@ import test from 'tape';
 import sinon from 'sinon';
 import addMilliseconds from 'date-fns/add_milliseconds';
 import subMilliseconds from 'date-fns/sub_milliseconds';
-import {
-  chat,
-  dequeueChatMessages,
-  __RewireAPI__ as RewireAPI,
-} from '../../src/utils/interface';
+
+import { chat, dequeueChatMessages, __RewireAPI__ as ChatRewireAPI } from '../chat';
 
 
 test('chat()', tester => {
   tester.test('queues messages in the correct order with back-to-back calls', t => {
     const mockQueue = [];
-    RewireAPI.__Rewire__('messageQueue', mockQueue);
+    ChatRewireAPI.__Rewire__('messageQueue', mockQueue);
 
     chat('one');
     chat('two');
@@ -20,18 +17,19 @@ test('chat()', tester => {
     chat('four');
 
     t.same(mockQueue, ['one', 'two', 'three', 'four']);
-    RewireAPI.__ResetDependency__('messageQueue');
+    ChatRewireAPI.__ResetDependency__('messageQueue');
     t.end();
   });
   tester.end();
 });
+
 
 test('dequeueChatMessages()', tester => {
   tester.test('chats first message of the program and dequeues the queue', t => {
     const mockQueue = ['one', 'two', 'three'];
     const mockEmit = sinon.spy();
 
-    RewireAPI.__Rewire__('messageQueue', mockQueue);
+    ChatRewireAPI.__Rewire__('messageQueue', mockQueue);
     global.tagpro = { socket: { emit: mockEmit } };
 
     dequeueChatMessages();
@@ -43,7 +41,7 @@ test('dequeueChatMessages()', tester => {
     t.same(mockQueue, ['two', 'three']);
 
     global.tagpro = undefined;
-    RewireAPI.__ResetDependency__('messageQueue');
+    ChatRewireAPI.__ResetDependency__('messageQueue');
     t.end();
   });
 
@@ -52,8 +50,8 @@ test('dequeueChatMessages()', tester => {
     const mockLastMessageTime = subMilliseconds(new Date(), 5000); // 5s in the past
     const mockEmit = sinon.spy();
 
-    RewireAPI.__Rewire__('messageQueue', mockQueue);
-    RewireAPI.__Rewire__('lastMessageTime', mockLastMessageTime);
+    ChatRewireAPI.__Rewire__('messageQueue', mockQueue);
+    ChatRewireAPI.__Rewire__('lastMessageTime', mockLastMessageTime);
     global.tagpro = { socket: { emit: mockEmit } };
 
     dequeueChatMessages();
@@ -65,8 +63,8 @@ test('dequeueChatMessages()', tester => {
     t.same(mockQueue, ['two', 'three']);
 
     global.tagpro = undefined;
-    RewireAPI.__ResetDependency__('messageQueue');
-    RewireAPI.__ResetDependency__('lastMessageTime');
+    ChatRewireAPI.__ResetDependency__('messageQueue');
+    ChatRewireAPI.__ResetDependency__('lastMessageTime');
     t.end();
   });
 
@@ -75,8 +73,8 @@ test('dequeueChatMessages()', tester => {
     const mockLastMessageTime = addMilliseconds(new Date(), 5000); // 5s in the future
     const mockEmit = sinon.spy();
 
-    RewireAPI.__Rewire__('messageQueue', mockQueue);
-    RewireAPI.__Rewire__('lastMessageTime', mockLastMessageTime);
+    ChatRewireAPI.__Rewire__('messageQueue', mockQueue);
+    ChatRewireAPI.__Rewire__('lastMessageTime', mockLastMessageTime);
     global.tagpro = { socket: { emit: mockEmit } };
 
     dequeueChatMessages();
@@ -85,8 +83,8 @@ test('dequeueChatMessages()', tester => {
     t.same(mockQueue, ['one', 'two', 'three']);
 
     global.tagpro = undefined;
-    RewireAPI.__ResetDependency__('messageQueue');
-    RewireAPI.__ResetDependency__('lastMessageTime');
+    ChatRewireAPI.__ResetDependency__('messageQueue');
+    ChatRewireAPI.__ResetDependency__('lastMessageTime');
     t.end();
   });
   tester.end();
